Finalize ThemeScreen with passing validation and CodeQL
Agent-Logs-Url: https://github.com/shelbeely/Epaper-Journal/sessions/0a900e5f-51d9-49ba-89f1-5770b721206f

+++ src/ui/ThemeScreen.cpp
@@ -263,7 +263,7 @@ void ThemeScreen::_render(const ThemeScreenData& data, int noteTop, int ratingWe
         uint16_t maxW = (tagX > textX + 6) ? (uint16_t)(tagX - textX - 6) : 40;
         uint16_t maxChars = (uint16_t)(maxW / _display.charAdvance(BODY_SCALE));
         auto lines = _wrapWords(n.text, maxChars, 2);
-        for (uint8_t li = 0; li < lines.size(); li++) {
+        for (size_t li = 0; li < lines.size(); li++) {
             _display.drawText(fb, textX, y + 2 + li * _display.lineHeight(BODY_SCALE),
                               lines[li].c_str(), false, BODY_SCALE);
         }
